+++ app/core/config.py
@@ -54,6 +54,7 @@ class Settings(BaseSettings):
     KLAS_LECTURE_URL: str
     KLAS_RECORDED_LECTURE_URL: str
     KLAS_COURSE_INFO_URL: str
+    KLAS_TEAM_PROJECT_URL: str
     KLAS_PROFILE_URL: str
     KLAS_STUDENT_INFO_URL: str
     KLAS_STUDENT_INFO_API_URL: str